Implemented changes from develop-merge and moved Effects into sub-package
