+++ app/src/main/java/at/vunfer/openrealms/model/Card.java
@@ -5,9 +5,9 @@ import java.util.List;
 
 public class Card {
     private static final String TAG = "Card";
-    private String name;
-    private int cost;
-    private List<Effect> effects;
+    private final String name;
+    private final int cost;
+    private final List<Effect> effects;
 
     public Card(String name, int cost, List<Effect> effects) throws IllegalArgumentException {
         if (name == null || name.isEmpty()) {
